// lib/tebex/generated/schema.ts
export interface components {
  schemas: {
    Webstore: {
      /** @example 1 */
      id?: number;
      /**
       * @description Store description in HTML format
       * @example <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Mauris fermentum eget tellus et cursus. Pellentesque vel vehicula velit, sit amet tincidunt diam. Aliquam sit amet fermentum lorem.</p>
       */
      description?: string;
      /**
       * @description Name of the store
       * @example Minecraft Store
       */
      name?: string;
      /**
       * Format: uri
       * @description URL of the webstore
       * @example https://example.tebex.io
       */
      webstore_url?: string;
      /**
       * @description Currency used in the store
       * @example USD
       */
      currency?: string;
      /**
       * @description Language of the store
       * @example en
       */
      lang?: string;
      /**
       * Format: uri
       * @description URL of the store's logo
       */
      logo?: string | null;
      /**
       * @description Platform type for the store
       * @example Minecraft: Java Edition
       */
      platform_type?: string;
      /** @example minecraft */
      platform_type_id?: string;
      /**
       * @description True if the store is disabled.
       * @example false
       */
      disabled?: boolean;
      /**
       * Format: date-time
       * @description The date and time when the store was created
       * @example 2023-05-25T09:49:37+00:00
       */
      created_at?: string;
    };
// ...
    Category: {
      /** @example 2678660 */
      id?: number;
      /**
       * @description User-friendly name of the category
       * @example Test
       */
      name?: string;
      /** @description Short identifier slug for the category */
      slug?: string | null;
      /**
       * Format: Category
       * @description Parent category, if applicable
       */
      parent?: Record<string, never> | null;
      /**
       * @description True if this is a tiered category
       * @example false
       */
      tiered?: boolean;
      /** @description If this is a tiered category and the usernameId is provided, this will be the active tier information for this category. */
      active_tier?: components["schemas"]["Tier"];
      /** @description HTML description of the category */
      description?: string;
      /** @description Packages within the category. For dynamic categories, packages use the `DynamicPackage` shape and are only populated when a `basketIdent` is provided. */
      packages?: components["schemas"]["Package"][] | null;
      /**
       * @description The numeric order in which to display the category.
       * @example 1
       */
      order?: number;
      /**
       * @description How the category should be displayed
       * @enum {string}
       */
      display_type?: "list" | "grid";
      /**
       * Format: uri
       * @description URL of the category image, if set.
       * @example null
       */
      image_url?: string | null;
      /**
       * @description True if this is a dynamic category. Dynamic categories are populated with custom packages per-basket using the Add Dynamic Packages endpoint, and their packages can only be fetched by providing a `basketIdent`.
       * @example false
       */
      dynamic?: boolean;
    };
    PackageResponse: {
      data?: components["schemas"]["Package"][];
    };
    Package: {
      /** @example 6276316 */
      id?: number;
      /** @example test package 101 */
      name?: string;
      /** @example <p>this is just a test. only a test. nothing more.</p> */
      description?: string;
      /** @example null */
      image?: string | null;
      /** @example single */
      type?: string;
      category?: {
        /** @example 2678660 */
        id?: number;
        /** @example Packages */
        name?: string;
      };
      /**
       * Format: float
       * @example 5
       */
      base_price?: number;
      /**
       * Format: float
       * @example 0
       */
      sales_tax?: number;
      /**
       * Format: float
       * @example 5
       */
      total_price?: number;
      /** @example USD */
      currency?: string;
      /**
       * Format: float
       * @description If this package is part of a tiered category, this is the difference on upgrade pricing from the current active tier.
       */
      prorate_price?: number | null;
      /**
       * Format: float
       * @example 0
       */
      discount?: number;
      /** @example false */
      disable_quantity?: boolean;
      /** @example false */
      disable_gifting?: boolean;
      /**
       * Format: date-time
       * @example null
       */
      expiration_date?: string | null;
      media?: components["schemas"]["PackageMedia"][];
      order?: number;
      slug?: string;
      user_limit?: number;
      creator_meta_data?: Record<string, never>;
      options?: unknown[];
      variables?: unknown[];
      /**
       * Format: date-time
       * @example 2024-05-21T18:09:17+00:00
       */
      created_at?: string;
      /**
       * Format: date-time
       * @example 2024-05-21T18:09:17+00:00
       */
      updated_at?: string;
    };
// ...
    PackageMedia: {
      /**
       * @example image
       * @enum {string}
       */
      type?: "video" | "image";
      /**
       * @description A file name for the media, if applicable.
       * @example e875d1010b2edb7f8c376728ab9dedfffe5bd012.jpg
       */
      name?: string | null;
      /**
       * @description The URL of the package media.
       * @example https://dunb17ur4ymx4.cloudfront.net/packages/images/e875d1010b2edb7f8c376728ab9dedfffe5bd012.jpg
       */
      url?: string;
      /**
       * @description True if this media should be featured more prominently than others.
       * @example false
       */
      featured?: boolean;
      /**
       * @description True if this is the primary media for the package.
       * @example true
       */
      primary?: boolean;
    };
// ...
  };
  responses: never;
  parameters: {
    /** @description The basket identifier. */
    basketIdent: string;
  };
  requestBodies: never;
  headers: never;
  pathItems: never;
}